Clarify "AgX" approach, remove corresponding curve
TS: "[the curves] do not matter in the least, [...] no degree of
precision [in them] can afford much utility"
Everyone else(?): "it's not AgX if it's not this exact curve"

Removing that ugly polynomial approximation to the original curve seems
appropriate to drive home the point.
TODO: find a simpler configurable curve in the same style.